Move fakeHostClient into testutils package
